requirements.txt
flask
instagrapi
numpy
pillow
psycopg2-binary
requests